solved all exercises

+++ week03/ex2.c
@@ -11,16 +11,16 @@ float distance(Point A, Point B){
 }
 
 float area(Point A, Point B, Point C){
-    return 1.0 / 2.0 * fabs(A.x*B.y - B.x*A.y + B.x*C.y - C.x*B.y + C.x*A.y - A.x*C.y);
+    return 1.0f / 2.0f * fabsf(A.x*B.y - B.x*A.y + B.x*C.y - C.x*B.y + C.x*A.y - A.x*C.y);
 }
 
 int main(void){
     Point A, B, C;
-    A.x = 2.5;
+    A.x = 2.5f;
     B.x = 1;
     C.x = 10;
     A.y = 6;
-    B.y = 2.2;
+    B.y = 2.2f;
     C.y = 6;
 
     float f = distance(A, B);
